blas rearrangements
